Add an option for shorter URLs
The method is to create the same long sha hash as before
bit start with the first four characters and check against
the existing ones in the database to see if it's unique. And
while there are duplicates add one character untill we're
back on the full hash string again.